Accidentally commit secrets to repo
This commit should have included a `.gitignore` file that excluded `secrets.json` from the repo.

Unfortunately, it didn't, so there are now secrets in the repo.

If this was the most recent commit, it would be easier to correct using `git commit --amend` and force push.

But unfortunately, we are going to add some more commits on top of it!