Failed attempt to exclude secrets from repo
At this point, I needed to update `secrets.json` and realized that I had accidentally included it in the repo.

I tried to correct this by adding `secrets.json` to my `.gitignore`, but as you can see, that didn't work.

Instead, this commit has the added `.gitignore` file, but still includes not only the previous secrets file, but also my changes to it.